Renamed APriori X,Y, and Z and APriori XSigma, YSigma, and ZSigma Filter files to be consistent with their counterparts for latitude, longitude, and radius.
