+++ isis/src/qisis/objs/CnetEditorWidget/APrioriXSigmaFilter.h
@@ -1,5 +1,13 @@
-#ifndef APrioriZFilter_H
-#define APrioriZFilter_H
+#ifndef APrioriXSigmaFilter_H
+#define APrioriXSigmaFilter_H
+
+/** This is free and unencumbered software released into the public domain.
+
+The authors of ISIS do not claim copyright on the contents of this file.
+For more details about the LICENSE terms and the AUTHORS, you will
+find files of those names at the top level of this repository. **/
+
+/* SPDX-License-Identifier: CC0-1.0 */
 
 #include "AbstractNumberFilter.h"
 
@@ -12,29 +20,29 @@ namespace Isis {
   class ControlPoint;
 
   /**
-   * @brief Allows filtering by a priori surface point Z
+   * @brief Allows filtering by a priori surface point X sigma
    *
    * This class allows the user to filter control points and control measures
-   * by a priori surface point Z. This allows the user to make a list
-   * of control points that are less than or greater than a certain a priori
-   * surface point Z.
+   * by a priori surface point X sigma. This allows the user to make a
+   * list of control points that are less than or greater than a certain
+   * a priori surface point X sigma.
    *
    * @author 2019-07-26 Ken Edmundson
    *
    * @internal
    *   @history 2019-07-26 Ken Edmundson - Original version based off of the
-   *                           APrioriRadiusFilter class.
+   *                           APrioriLatitudeSigmaFilter class.
    *   @history 2019-11-22 Ken Edmundson - Minor text corrections to returned QString in
    *                           getImageDescription() and getPointDescription() methods.
    */
-  class APrioriZFilter : public AbstractNumberFilter {
+  class APrioriXSigmaFilter : public AbstractNumberFilter {
       Q_OBJECT
 
     public:
-      APrioriZFilter(AbstractFilter::FilterEffectivenessFlag flag,
+      APrioriXSigmaFilter(AbstractFilter::FilterEffectivenessFlag flag,
             int minimumForSuccess = -1);
-      APrioriZFilter(const APrioriZFilter &other);
-      virtual ~APrioriZFilter();
+      APrioriXSigmaFilter(const APrioriXSigmaFilter &other);
+      virtual ~APrioriXSigmaFilter();
 
       bool evaluate(const QPair<QString, ControlNet *> *) const;
       bool evaluate(const ControlPoint *) const;
