+++ isis/src/qisis/objs/CnetEditorWidget/APrioriZSigmaFilter.h
@@ -1,5 +1,13 @@
-#ifndef APrioriYFilter_H
-#define APrioriYFilter_H
+#ifndef APrioriZSigmaFilter_H
+#define APrioriZSigmaFilter_H
+
+/** This is free and unencumbered software released into the public domain.
+
+The authors of ISIS do not claim copyright on the contents of this file.
+For more details about the LICENSE terms and the AUTHORS, you will
+find files of those names at the top level of this repository. **/
+
+/* SPDX-License-Identifier: CC0-1.0 */
 
 #include "AbstractNumberFilter.h"
 
@@ -12,29 +20,29 @@ namespace Isis {
   class ControlPoint;
 
   /**
-   * @brief Allows filtering by a priori surface point Y
+   * @brief Allows filtering by a priori surface point Z sigma
    *
    * This class allows the user to filter control points and control measures
-   * by a priori surface point Y. This allows the user to make a list
-   * of control points that are less than or greater than a certain a priori
-   * surface point Y.
+   * by a priori surface point Z sigma. This allows the user to make a
+   * list of control points that are less than or greater than a certain
+   * a priori surface point Z sigma.
    *
    * @author 2019-07-26 Ken Edmundson
    *
    * @internal
    *   @history 2019-07-26 Ken Edmundson - Original version based off of the
-   *                           APrioriLongitudeFilter class.
+   *                           APrioriRadiusSigmaFilter class.
    *   @history 2019-11-22 Ken Edmundson - Minor text corrections to returned QString in
    *                           getImageDescription() and getPointDescription() methods.
    */
-  class APrioriYFilter : public AbstractNumberFilter {
+  class APrioriZSigmaFilter : public AbstractNumberFilter {
       Q_OBJECT
 
     public:
-      APrioriYFilter(AbstractFilter::FilterEffectivenessFlag flag,
+      APrioriZSigmaFilter(AbstractFilter::FilterEffectivenessFlag flag,
             int minimumForSuccess = -1);
-      APrioriYFilter(const APrioriYFilter &other);
-      virtual ~APrioriYFilter();
+      APrioriZSigmaFilter(const APrioriZSigmaFilter &other);
+      virtual ~APrioriZSigmaFilter();
 
       bool evaluate(const QPair<QString, ControlNet *> *) const;
       bool evaluate(const ControlPoint *) const;
